docs: add section for contributors to docs
This commit removes top-level headings already enforced by `title:` and includes a complete Contributing to the Docs page intended to help guide new external and internal contributors to our documentation who may or may not be Git-adept. It is also intended to serve as supporting documentation for any technical writer role we may onboard in the future.